Power BI
To understand the data that I uploaded in an incomprehensible tabular form, I visualize them in power bi to make a conclusion for myself and show the guide my decision and why I made it/I want to do it. Before I visualize the data on the chart, I clean them, if necessary, combine and merge tables (Power query). Then I tie it up.